+++ src/main/java/com/example/BuildPC/repository/AdminRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdminRepository extends JpaRepository<User, Integer> {
+public interface AdminRepository extends JpaRepository<User, Long> {
 }
